requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
aiofiles
numpy
tqdm
pyyaml
imageio
tifffile
python-multipart
future
pynwb
psutil
nbstripout
snakemake
# git+https://github.com/ShogoAkiyama/snakemake@main#egg=snakemake
jinja2==3.0.3
firebase_admin
pydantic[email]
pydantic[dotenv]
pyrebase4==4.6.0
requests-toolbelt==0.10.1
python-jose[cryptography]
nibabel==5.1.0
matplotlib==3.7.1
